Added dock functionality + menubar
Added dock functionality by putting a QMainWindow into the QDialog of the widget. Unsure if this is the best way to go.

Added a menubar with options to show dockable widgets if they are closed

Tried to implement a timer to solve the redraw problem but it did not work.

TODO
-Save and load dock window state
-setMaximum on QSlider sometimes(?) fails when passed a float
-address redraw problem